backend/pyproject.toml
[tool.poetry]
name = "census_api"
version = "0.1.0"
description = ""
authors = ["Guillaume Mazoyer <oss@mazoyer.eu>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
gunicorn = "^22.0.0"
fastapi = "^0.111.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
httpx = "^0.27.0"
psycopg = { extras = ["binary"], version = "^3.1.19" }
alembic = "^1.13.1"
sqlmodel = "^0.0.19"
tenacity = "^8.4.1"
emoji-country-flag = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.21.1"
pytest-httpx = "^0.30.0"
aiosqlite = "^0.20.0"
mypy = "^1.10.0"
ruff = "^0.5.0"

[build-system]
requires = ["poetry>=1.8.3"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = "census_api/tests"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "A",     # flake8-builtins
    "ARG",   # flake8-unused-arguments
    "ASYNC", # flake8-async
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "E",     # pycodestyle errors
    "ERA",   # eradicate
    "EXE",   # flake8-executable
    "F",     # pyflakes
    "FA",    # flake8-future-annotations
    "FBT",   # flake8-boolean-trap
    "I",     # isort
    "ICN",   # flake8-import-conventions
    "INP",   # flake8-no-pep420
    "ISC",   # flake8-implicit-str-concat
    "N",     # pep8-naming
    "PIE",   # flake8-pie
    "PL",    # pylint
    "PT",    # flake8-pytest-style
    "PTH",   # flake8-use-pathlib
    "PYI",   # flake8-pyi
    "Q",     # flake8-quotes
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "RUF",   # ruff specific rules
    "SIM",   # flake8-simplify
    "SLF",   # flake8-self
    "SLOT",  # flake8-slots
    "T10",   # flake8-debugger
    "T20",   # flake8-print
    "TCH",   # flake8-type-checking
    "UP",    # pyupgrade
    "W",     # pycodestyle warnings
    "YTT",   # flake8-2020
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.ruff.lint.per-file-ignores]
"census_api/tests/**" = ["ARG001"]
